Refactored macro-related parts of Utils into MacroUtils to make a hard division between macro and non-macro classes and therefore class-loading.
